Final Version
Node.JS project.
This program analyzes all files that have a specific extension given in the command line as a parameter and looks for a specific string (also given in the command line as a parameter) and returns the filename and path. If there are no parameters a message will be prompt to the user the correct way to input information and if there are no results for the search it will be also prompt to the user.